refactor: migrate clamp components to SFCs
